fix: update utils, check duplicate usernames
